pyproject.toml
[tool.poetry]
name = "ai-coder"
version = "0.1.0"
description = ""
authors = ["Hao Qian <2035305+chandlerq@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.scripts]
ai_coder = "ai_coder.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.11"
langchain-openai = "^0.0.8"
astor = "^0.8.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
black = "^24.3.0"
ruff = "^0.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
